Fixed pthread bug when exiting

+++ neo/sys/sys_public.h
@@ -482,7 +482,7 @@ void			Sys_ShutdownNetworking( void );
 ==============================================================
 */
 
-typedef unsigned int (*xthread_t)( void * );
+typedef size_t (*xthread_t)( void * );
 
 typedef enum {
 	THREAD_NORMAL,
@@ -492,7 +492,7 @@ typedef enum {
 
 typedef struct {
 	const char *	name;
-	int				threadHandle;
+	size_t				threadHandle;
 	unsigned long	threadId;
 } xthreadInfo;
